requirements.txt
PyYAML==6.0
requests==2.28.2
tabulate==0.9.0
wcwidth==0.2.5
beautifulsoup4==4.10.0
